Modify code to meet requirements
