+++ main.rb
@@ -1,4 +1,5 @@
 require_relative 'app'
+require_relative 'logic'
 def main
   books = []
   people = []
@@ -22,23 +23,5 @@ def display_menu
   puts '6. List rentals for a person'
   puts '7. Quit'
   print 'Select an option: '
-end
-
-def handle_option(option, books, people, rentals)
-  option_actions = {
-    1 => -> { list_all_books(books) },
-    2 => -> { list_all_people(people) },
-    3 => -> { create_person(people) },
-    4 => -> { create_book(books) },
-    5 => -> { create_rental(people, books, rentals) },
-    6 => -> { list_rentals(rentals) },
-    7 => -> { quit }
-  }
-  action = option_actions[option]
-  if action
-    action.call
-  else
-    puts 'Invalid option!'
-  end
 end
 main
